testing file
A testing file has been made for viewing the changes that had to be done in the code. If you all approve, we can copy paste testing.html to the main directory.

I still have to alter the font size of the file, other than that if you guys have any suggestions, just drop them in the whatsapp group